refactor(app): migrate the image hook to a shared namepsace
