+++ app/src/organisms/ODD/RunningProtocol/ImageGalleryList/GalleryListItem.tsx
@@ -5,9 +5,8 @@ import { ListItem, StyledText } from '@opentrons/components'
 import { SmallButton } from '/app/atoms/buttons'
 import { Skeleton } from '/app/atoms/Skeleton'
 import { useImageGalleryData } from '/app/local-resources/dataFiles/hooks/useImageGalleryData'
-// eslint-disable-next-line opentrons/no-imports-across-applications
-import { useImage } from '/app/organisms/Desktop/Devices/ProtocolRun/ProtocolRunCamera/ImageGalleryContainer/GalleryItemCard'
 import { handleCameraPhotoModal } from '/app/organisms/ODD/RunningProtocol/ImageGalleryList/CameraPhotoModal'
+import { useImage } from '/app/resources/dataFiles/useImage'
 
 import styles from './gallery.module.css'
